// src/modules/inventory/reports/yarn-store/knit-dyeing-price-overview-report/knit-dyeing-price-overview-report-type.ts
export type IKnittingDyeingPriceOverviewReport = {
  BUYER_ID: number;
  BUYER: string;
  PO_ID: number;
  PO_NO: string;
  STYLE_ID: number;
  STYLENO: string;
  STYLENAME: string;
  ITEMTYPE: string;
  COLOR_ID: number;
  COLOR: string;
  FABRIC_PART_ID: number;
  FABRIC_PART: string;
  FABRIC_ID: number;
  FABRIC: string;
  PO_STYLE_WISE_ORDER_QTY: string;
  DELIVERYDATE: Date;
  YARN_COUNT: string;
  GSM: string;
  YARN_ISSUE_QTY: number;
  GREY_RCV_QTY: number;
  KNIT_RATE: number;
  KNITTING_AMOUNT: number;
  GREY_QUANTITY: number;
  FINISH_QUANTITY: number;
  SHADE_PERCENT: string;
  DYEING_RATE: string;
  DYEING_AMOUNT: number;
};